Refactor to remove Delete because a few of the customer apis don't have it
